Restructure your existing UI code
Organized packages and files
