+++ app/src/main/java/com/kodeco/android/countryinfo/ui/screens/countryDetails/CountryDetailsScreen.kt
@@ -1,4 +1,4 @@
-package com.kodeco.android.countryinfo.ui.components
+package com.kodeco.android.countryinfo.ui.screens.countryDetails
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxSize
@@ -23,6 +23,7 @@ import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.kodeco.android.countryinfo.flow.Flows
 import com.kodeco.android.countryinfo.models.Country
+import com.kodeco.android.countryinfo.ui.components.sampleCountry
 
 @Composable
 fun CountryDetailsScreen(country: Country, onBackPress: () -> Unit) {
