+++ app/src/main/java/com/kodeco/android/countryinfo/ui/screens/countryInfo/CountryInfoScreen.kt
@@ -1,4 +1,4 @@
-package com.kodeco.android.countryinfo.ui.components
+package com.kodeco.android.countryinfo.ui.screens.countryInfo
 
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
@@ -13,6 +13,9 @@ import com.kodeco.android.countryinfo.flow.Flows
 import com.kodeco.android.countryinfo.models.Country
 import com.kodeco.android.countryinfo.models.CountryFlags
 import com.kodeco.android.countryinfo.models.CountryName
+import com.kodeco.android.countryinfo.ui.components.CountryErrorScreen
+import com.kodeco.android.countryinfo.ui.components.CountryInfoList
+import com.kodeco.android.countryinfo.ui.components.Loading
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
