+++ app/src/main/java/com/kodeco/android/countryinfo/ui/screens/countryInfo/CountryInfoViewModel.kt
@@ -1,0 +1,7 @@
+package com.kodeco.android.countryinfo.ui.screens.countryInfo
+
+import androidx.lifecycle.ViewModel
+
+class CountryInfoViewModel: ViewModel() {
+
+}
